[API] Move a file under Extensions
